Moving api.ts to a new 'util' directory
